ci: speed up regular test workflow
Reuse expensive test runtimes, parallelize Javadoc reactor work, and run trusted builds on the 8-core runner.
